refactor: Add support for --no-cache flag in build_image function
The `build_image` function in `dockerbuilderhelper.py` has been updated to include support for the `--no-cache` flag. This flag can be passed as a parameter to the function, allowing the Docker image to be built without using the cache. This provides a way to ensure that the image is built from scratch, without any cached layers.