refactor: Add '-d' flag for detached mode in run_compose function
The `run_compose` function in `dockerbuilderhelper.py` has been updated to include the `-d` flag when running the `docker-compose up` command. This flag allows the containers to run in detached mode, meaning they will continue to run in the background even after the command has finished executing. This change improves the usability of the function by providing a more convenient way to start the containers.